Several Bug Fixes in Rendezvous, seems to be working ok.   Using a better SRNextHopStrategy which doesn't use NATted nodes   Return null if there is no way to get to the intended dest, this is important b/c it usually happens before the leafset has been built up, and just need to make these checks fail until the leafset is complete
Fixed issue in FileTransfer s.t. if there is a failure in the constructor of the FileReceipt, it doesn't leave broken state.
Fixed bug in OptionsFactory where it was assigning the wrong options
Better logging in several TLs
Fixed ScribeRegrTest to make it delay much longer for some tests
Better toString()s in some messages
Made SPNF more extensible


git-svn-id: https://svn.mpi-sws.org/DS/freepastry/trunk@4143 c21dacaf-8602-0410-aacd-8bc41df2cbe7